feat: #71 Move Components in cqdg for more clarity and easier refactoring
