+++ public/index.html
@@ -48,20 +48,5 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-    <!-- legacy Angular stuff -->
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.1.1/jquery.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/angular.js/1.4.14/angular.min.js"></script>
-    <script src="/gdc-ng-plugins.js"></script>
-    <script src="/gdc-gql.js"></script>
-    <script src="/gdc-templates.js"></script>
-    <script src="/gdc-config.js"></script>
-    <script src="/gdc-app-modified.js"></script>
-    <!-- END legacy Angular stuff -->
-
-    <script
-      language="JavaScript"
-      type="text/javascript"
-      src="https://static.cancer.gov/webanalytics/wa_gdc_pageload.js"
-    ></script>
   </body>
 </html>
