+++ src/packages/@cqdg/pages/FileRepository/CaseAggregations/CaseAggregations.relay.js
@@ -4,21 +4,20 @@ import React from 'react';
 import { graphql } from 'react-relay';
 import { compose, withPropsOnChange } from 'recompose';
 import { connect } from 'react-redux';
+import { parse } from 'query-string';
 import Query from '@ncigdc/modern_components/Query';
 import { parseFilterParam } from '@ncigdc/utils/uri';
 import withRouter from '@ncigdc/utils/withRouter';
-import { parse } from 'query-string';
-import {mapFilter} from "../../utils/filters";
-import {repoPageCaseToFileFiltersMapping} from "../../containers/RepositoryPage";
+import { repoPageFileToCaseFiltersMapping } from '@cqdg/pages/FileRepository/FilterMapping';
+import { mapFilter } from '@ncigdc/utils/filters';
 
 const entityType = 'Files';
-
 export default (Component: ReactClass<*>) =>
   compose(
     withRouter,
     withPropsOnChange(
       ['location'],
-      ({ location: { search }, defaultFilters = null }) => {
+      ({ defaultFilters = null, location: { search } }) => {
         const q = parse(search);
         const filters = parseFilterParam(q.filters, defaultFilters);
         return {
@@ -26,98 +25,102 @@ export default (Component: ReactClass<*>) =>
         };
       },
     ),
-    connect((state, props) => ({
+    connect((state) => ({
       userSelectedFacets: state.customFacets[entityType],
     })),
     withPropsOnChange(
       ['userSelectedFacets', 'filters'],
-      ({ userSelectedFacets, filters }) => {
+      ({ filters, userSelectedFacets }) => {
         return {
           variables: {
             filters,
-            repoFileCustomFacetFields: userSelectedFacets
+            repoCaseCustomFacetFields: userSelectedFacets
               .map(({ field }) => field)
               .join(','),
           },
         };
       },
     ),
-  )((props: Object) => {
-
-    if(props && props.variables && props.variables.filters){
-      mapFilter(props.variables.filters, repoPageCaseToFileFiltersMapping)
+  )((props) => {
+    if (props && props.variables && props.variables.filters) {
+      mapFilter(props.variables.filters, repoPageFileToCaseFiltersMapping);
     }
 
     return (
       <Query
-        parentProps={props}
-        minHeight={578}
-        variables={props.variables}
         Component={Component}
+        minHeight={578}
+        parentProps={props}
         query={graphql`
-          query FileAggregations_relayQuery(
+          query CaseAggregations_relayQuery(
             $filters: JSON
           ) {
             viewer {
-                File {
+              Case { 
                   aggregations(
                     filters: $filters
                     aggregations_filter_themselves: false
                   ) {
-                    data_category {
-                      buckets {
-                        doc_count
-                        key
-                      }
-                    }
-                    data_type {
-                      buckets {
-                        doc_count
-                        key
-                      }
-                    }
-                    is_harmonized{
+                    study__short_name_keyword{
                       buckets{
                         doc_count
                         key
-                        key_as_string
                       }
-                    }   
-                    experimental_strategy {
-                      buckets {
-                        doc_count
-                        key
-                      }
-                    }             
-                    file_format {
-                      buckets {
+                    }
+                    study__study_id_keyword{
+                      buckets{
                         doc_count
                         key
                       }
                     }
-                    data_access {
-                      buckets {
+                    study__domain{
+                      buckets{
                         doc_count
                         key
                       }
                     }
-                    platform {
-                      buckets {
+                    gender{
+                      buckets{
                         doc_count
                         key
                       }
                     }
-                    file_variant_class{
+                    ethnicity{
+                      buckets{
+                        doc_count
+                        key
+                      }
+                    }
+                    vital_status{
+                      buckets{
+                        doc_count
+                        key
+                      }
+                    }
+                    diagnoses__icd_category_keyword{
+                      buckets{
+                        doc_count
+                        key
+                      }
+                    }
+                    phenotypes__hpo_category_keyword{
+                      buckets{
+                        doc_count
+                        key
+                      }
+                    }
+                    phenotypes__hpo_term_keyword{
                       buckets{
                         doc_count
                         key
                       }
                     }
                   }
+                }
               }
-            }
           }
         `}
-      />
+        variables={props.variables}
+        />
     );
   });
